refactor: isolate migration maintenance
